+++ libs/cmcd/src/fromCmcdHeaders.ts
@@ -1,7 +1,7 @@
 import type { Cmcd } from './Cmcd.ts'
 import { CMCD_HEADER_FIELDS } from './CmcdHeaderField.ts'
 import { decodeCmcd } from './decodeCmcd.ts'
-
+import { ensureHeaders } from './ensureHeaders.ts'
 
 /**
  * Decode CMCD data from request headers.
@@ -16,12 +16,10 @@ import { decodeCmcd } from './decodeCmcd.ts'
  * {@includeCode ../test/fromCmcdHeaders.test.ts#example}
  */
 export function fromCmcdHeaders(headers: Record<string, string> | Headers): Cmcd {
-	if (!(headers instanceof Headers)) {
-		headers = new Headers(headers)
-	}
+	const h = ensureHeaders(headers)
 
 	return CMCD_HEADER_FIELDS.reduce((acc, key) => {
-		const value = headers.get(key)
+		const value = h.get(key)
 		return Object.assign(acc, decodeCmcd(value as string))
 	}, {} as Cmcd)
 }